+++ brewtils/auto_decorator.py
@@ -1,4 +1,6 @@
-import inspect
+from inspect import Parameter as InspectParameter  # noqa
+from inspect import signature
+
 
 from brewtils.models import Command, Parameter
 
@@ -27,9 +29,15 @@ class AutoDecorator:
                 _wrapped = getattr(client, func)
                 if not hasattr(_wrapped, "_command") and not func.startswith("__"):
                     # decorators.py will handle all of the markings
-                    if func.startswith("_"):
-                        _wrapped._command = Command(hidden=True)
-                    else:
-                        _wrapped._command = Command()
+                    has_kwargs = any(
+                        [
+                            True
+                            for p in signature(_wrapped).parameters.values()
+                            if p.kind == InspectParameter.VAR_KEYWORD
+                        ]
+                    )
+                    _wrapped._command = Command(
+                        hidden=func.startswith("_"), allow_any_kwargs=has_kwargs
+                    )
 
         return client