+++ brewtils/rest/system_client.py
@@ -356,11 +356,15 @@ class SystemClient(object):
                 _command=command_name,
                 _system_name=self._system.name,
                 _system_namespace=self._system.namespace,
-                _system_version=self._system.version,
+                _system_version=(
+                    self._version_constraint
+                    if self._version_constraint == "latest"
+                    else self._system.version
+                ),
                 _system_display=self._system.display_name,
                 _output_type=self._commands[command_name].output_type,
                 _instance_name=self._default_instance,
-                **kwargs
+                **kwargs,
             )
         else:
             raise AttributeError(
@@ -459,11 +463,16 @@ class SystemClient(object):
         """
 
         if self._version_constraint == "latest":
-            self._system = self._determine_latest(
-                self._easy_client.find_systems(
-                    name=self._system_name, namespace=self._system_namespace
-                )
+            systems = self._easy_client.find_systems(
+                name=self._system_name,
+                namespace=self._system_namespace,
+                filter_latest=True,
             )
+            if systems:
+                self._system = systems[0]
+            else:
+                self._system = None
+
         else:
             self._system = self._easy_client.find_unique_system(
                 name=self._system_name,
@@ -510,6 +519,29 @@ class SystemClient(object):
         if raise_on_error and request.status == "ERROR":
             raise RequestFailedError(request)
 
+        # Support cross-server parent/child requests. Add parent if request has different host.
+        if (
+            request.parent is None
+            and brewtils.plugin.CONFIG
+            and (
+                brewtils.plugin.CONFIG.bg_host.upper()
+                != self._easy_client.client.bg_host.upper()
+                or brewtils.plugin.CONFIG.bg_port != self._easy_client.client.bg_port
+                or brewtils.plugin.CONFIG.bg_url_prefix
+                != self._easy_client.client.bg_prefix
+            )
+        ):
+            request.parent = getattr(
+                brewtils.plugin.request_context, "current_request", None
+            )
+            request.has_parent = request.parent is not None
+            ec = EasyClient(
+                bg_host=brewtils.plugin.CONFIG.bg_host,
+                bg_port=brewtils.plugin.CONFIG.bg_port,
+                bg_url_prefix=brewtils.plugin.CONFIG.bg_url_prefix,
+            )
+            ec.put_request(request)
+
         return request
 
     def _get_parent_for_request(self):
@@ -526,8 +558,7 @@ class SystemClient(object):
             self._logger.warning(
                 "A parent request was found, but the destination beer-garden "
                 "appears to be different than the beer-garden to which this plugin "
-                "is assigned. Cross-server parent/child requests are not supported "
-                "at this time. Removing the parent context so the request doesn't fail."
+                "is assigned. Removing the parent context so the request doesn't fail."
             )
             return None
 
@@ -604,7 +635,10 @@ class SystemClient(object):
             return request.parameters
 
         return self._resolver.resolve(
-            request.parameters, self._commands[command].parameters, upload=True
+            request.parameters,
+            self._commands[command].parameters,
+            upload=True,
+            allow_any_parameter=self._commands[command].allow_any_kwargs,
         )
 
     @staticmethod